Stupid CSS bug
Not letting me fucking use transform or margin/padding when Position: FIxed or Position: Sticky is used.

There is no way to scroll to an element with this. I can't test any text spacing with this bug. Unless I can find a solution, I'll have to implement a JS script that stickies for me.
I am very mad